+++ redclip/src/main/java/com/kh/redclip/member/model/service/MemberService.java
@@ -70,7 +70,7 @@ public interface MemberService {
 
 	int changePw(Member member);
 
-	
+
